트랜잭션 처리 리팩토링
defer만 따로 함수로 만들었더니 `ORA-01453: SET TRANSACTION must be first statement of transaction` 에러가 발생
커넥션 풀에 `READ WRITE 옵션`을 지정하는 `BeginTxWithMode()`함수를 작성하여 `SET TRANSACTION` 에러가 insert/update/delete작업에서는 발생하지 않도록 함.
