+++ csm-api/service/service_compare.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"csm-api/entity"
 	"csm-api/store"
+	"csm-api/txutil"
 	"csm-api/utils"
 	"strconv"
 	"strings"
@@ -251,12 +252,12 @@ func (s *ServiceCompare) GetCompareList(ctx context.Context, compare entity.Comp
 
 // 근로자 비교 반영
 func (s *ServiceCompare) ModifyWorkerCompareApply(ctx context.Context, workers entity.WorkerDailys) (err error) {
-	tx, err := s.SafeTDB.BeginTx(ctx, nil)
+	tx, err := txutil.BeginTxWithMode(ctx, s.SafeTDB, false)
 	if err != nil {
 		return utils.CustomErrorf(err)
 	}
 
-	defer utils.DeferTx(tx, &err)
+	defer txutil.DeferTx(tx, &err)
 
 	// 근로자 정보: IRIS_WORKER_SET
 	// 선택한 프로젝트로 수정
