+++ csm-api/service/service_notice.go
@@ -5,6 +5,7 @@ import (
 	"csm-api/auth"
 	"csm-api/entity"
 	"csm-api/store"
+	"csm-api/txutil"
 	"csm-api/utils"
 	"github.com/guregu/null"
 )
@@ -90,12 +91,12 @@ func (s *ServiceNotice) GetNoticeListCount(ctx context.Context, search entity.No
 // @param
 // - notice entity.Notice: JNO, TITLE, CONTENT, SHOW_YN, PERIOD_CODE, REG_UNO, REG_USER
 func (s *ServiceNotice) AddNotice(ctx context.Context, notice entity.Notice) (err error) {
-	tx, err := s.SafeTDB.BeginTx(ctx, nil)
+	tx, err := txutil.BeginTxWithMode(ctx, s.SafeTDB, false)
 	if err != nil {
 		return utils.CustomErrorf(err)
 	}
 
-	defer utils.DeferTx(tx, &err)
+	defer txutil.DeferTx(tx, &err)
 
 	if err = s.Store.AddNotice(ctx, tx, notice); err != nil {
 		return utils.CustomErrorf(err)
@@ -108,12 +109,12 @@ func (s *ServiceNotice) AddNotice(ctx context.Context, notice entity.Notice) (er
 // @param
 // -notice entity.Notice: IDX, JNO, TITLE, CONTENT, SHOW_YN, MOD_UNO, MOD_USER
 func (s *ServiceNotice) ModifyNotice(ctx context.Context, notice entity.Notice) (err error) {
-	tx, err := s.SafeTDB.BeginTx(ctx, nil)
+	tx, err := txutil.BeginTxWithMode(ctx, s.SafeTDB, false)
 	if err != nil {
 		return utils.CustomErrorf(err)
 	}
 
-	defer utils.DeferTx(tx, &err)
+	defer txutil.DeferTx(tx, &err)
 
 	if err = s.Store.ModifyNotice(ctx, tx, notice); err != nil {
 		return utils.CustomErrorf(err)
@@ -126,12 +127,12 @@ func (s *ServiceNotice) ModifyNotice(ctx context.Context, notice entity.Notice) 
 // @param
 // - IDX: 공지사항 인덱스
 func (s *ServiceNotice) RemoveNotice(ctx context.Context, idx null.Int) (err error) {
-	tx, err := s.SafeTDB.BeginTx(ctx, nil)
+	tx, err := txutil.BeginTxWithMode(ctx, s.SafeTDB, false)
 	if err != nil {
 		return utils.CustomErrorf(err)
 	}
 
-	defer utils.DeferTx(tx, &err)
+	defer txutil.DeferTx(tx, &err)
 
 	if err = s.Store.RemoveNotice(ctx, tx, idx); err != nil {
 		return utils.CustomErrorf(err)
