+++ csm-api/service/service_project.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"csm-api/entity"
 	"csm-api/store"
+	"csm-api/txutil"
 	"csm-api/utils"
 	"database/sql"
 	"fmt"
@@ -297,12 +298,12 @@ func (s *ServiceProject) GetProjectBySite(ctx context.Context, sno int64) (entit
 // @param
 // -
 func (s *ServiceProject) AddProject(ctx context.Context, project entity.ReqProject) (err error) {
-	tx, err := s.SafeTDB.BeginTx(ctx, nil)
+	tx, err := txutil.BeginTxWithMode(ctx, s.SafeTDB, false)
 	if err != nil {
 		return utils.CustomErrorf(err)
 	}
 
-	defer utils.DeferTx(tx, &err)
+	defer txutil.DeferTx(tx, &err)
 
 	err = s.Store.AddProject(ctx, tx, project)
 	if err != nil {
@@ -315,12 +316,12 @@ func (s *ServiceProject) AddProject(ctx context.Context, project entity.ReqProje
 // @param
 // -
 func (s *ServiceProject) ModifyDefaultProject(ctx context.Context, project entity.ReqProject) (err error) {
-	tx, err := s.SafeTDB.BeginTx(ctx, nil)
+	tx, err := txutil.BeginTxWithMode(ctx, s.SafeTDB, false)
 	if err != nil {
 		return utils.CustomErrorf(err)
 	}
 
-	defer utils.DeferTx(tx, &err)
+	defer txutil.DeferTx(tx, &err)
 
 	err = s.Store.ModifyDefaultProject(ctx, tx, project)
 	if err != nil {
@@ -333,12 +334,12 @@ func (s *ServiceProject) ModifyDefaultProject(ctx context.Context, project entit
 // @param
 // -
 func (s *ServiceProject) ModifyUseProject(ctx context.Context, project entity.ReqProject) (err error) {
-	tx, err := s.SafeTDB.BeginTx(ctx, nil)
+	tx, err := txutil.BeginTxWithMode(ctx, s.SafeTDB, false)
 	if err != nil {
 		return utils.CustomErrorf(err)
 	}
 
-	defer utils.DeferTx(tx, &err)
+	defer txutil.DeferTx(tx, &err)
 
 	err = s.Store.ModifyUseProject(ctx, tx, project)
 	if err != nil {
@@ -351,12 +352,12 @@ func (s *ServiceProject) ModifyUseProject(ctx context.Context, project entity.Re
 // @param
 // -
 func (s *ServiceProject) RemoveProject(ctx context.Context, sno int64, jno int64) (err error) {
-	tx, err := s.SafeTDB.BeginTx(ctx, nil)
+	tx, err := txutil.BeginTxWithMode(ctx, s.SafeTDB, false)
 	if err != nil {
 		return utils.CustomErrorf(err)
 	}
 
-	defer utils.DeferTx(tx, &err)
+	defer txutil.DeferTx(tx, &err)
 
 	err = s.Store.RemoveProject(ctx, tx, sno, jno)
 	if err != nil {
