+++ csm-api/service/service_project_daily.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"csm-api/entity"
 	"csm-api/store"
+	"csm-api/txutil"
 	"csm-api/utils"
 )
 
@@ -24,12 +25,12 @@ func (s *ServiceProjectDaily) GetDailyJobList(ctx context.Context, jno int64, ta
 
 // 작업내용 추가
 func (s *ServiceProjectDaily) AddDailyJob(ctx context.Context, project entity.ProjectDailys) (err error) {
-	tx, err := s.SafeTDB.BeginTx(ctx, nil)
+	tx, err := txutil.BeginTxWithMode(ctx, s.SafeTDB, false)
 	if err != nil {
 		return utils.CustomErrorf(err)
 	}
 
-	defer utils.DeferTx(tx, &err)
+	defer txutil.DeferTx(tx, &err)
 
 	if err = s.Store.AddDailyJob(ctx, tx, project); err != nil {
 		return utils.CustomErrorf(err)
@@ -39,12 +40,12 @@ func (s *ServiceProjectDaily) AddDailyJob(ctx context.Context, project entity.Pr
 
 // 작업내용 수정
 func (s *ServiceProjectDaily) ModifyDailyJob(ctx context.Context, project entity.ProjectDaily) (err error) {
-	tx, err := s.SafeTDB.BeginTx(ctx, nil)
+	tx, err := txutil.BeginTxWithMode(ctx, s.SafeTDB, false)
 	if err != nil {
 		return utils.CustomErrorf(err)
 	}
 
-	defer utils.DeferTx(tx, &err)
+	defer txutil.DeferTx(tx, &err)
 
 	if err = s.Store.ModifyDailyJob(ctx, tx, project); err != nil {
 		return utils.CustomErrorf(err)
@@ -54,12 +55,12 @@ func (s *ServiceProjectDaily) ModifyDailyJob(ctx context.Context, project entity
 
 // 작업내용 삭제
 func (s *ServiceProjectDaily) RemoveDailyJob(ctx context.Context, idx int64) (err error) {
-	tx, err := s.SafeTDB.BeginTx(ctx, nil)
+	tx, err := txutil.BeginTxWithMode(ctx, s.SafeTDB, false)
 	if err != nil {
 		return utils.CustomErrorf(err)
 	}
 
-	defer utils.DeferTx(tx, &err)
+	defer txutil.DeferTx(tx, &err)
 
 	if err = s.Store.RemoveDailyJob(ctx, tx, idx); err != nil {
 		return utils.CustomErrorf(err)
