+++ csm-api/service/service_schedule.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"csm-api/entity"
 	"csm-api/store"
+	"csm-api/txutil"
 	"csm-api/utils"
 )
 
@@ -29,12 +30,12 @@ func (s *ServiceSchedule) GetRestScheduleList(ctx context.Context, jno int64, ye
 // @param
 // -
 func (s *ServiceSchedule) AddRestSchedule(ctx context.Context, schedule entity.RestSchedules) (err error) {
-	tx, err := s.SafeTDB.BeginTx(ctx, nil)
+	tx, err := txutil.BeginTxWithMode(ctx, s.SafeTDB, false)
 	if err != nil {
 		return utils.CustomErrorf(err)
 	}
 
-	defer utils.DeferTx(tx, &err)
+	defer txutil.DeferTx(tx, &err)
 
 	err = s.Store.AddRestSchedule(ctx, tx, schedule)
 	if err != nil {
@@ -48,12 +49,12 @@ func (s *ServiceSchedule) AddRestSchedule(ctx context.Context, schedule entity.R
 // @param
 // -
 func (s *ServiceSchedule) ModifyRestSchedule(ctx context.Context, schedule entity.RestSchedule) (err error) {
-	tx, err := s.SafeTDB.BeginTx(ctx, nil)
+	tx, err := txutil.BeginTxWithMode(ctx, s.SafeTDB, false)
 	if err != nil {
 		return utils.CustomErrorf(err)
 	}
 
-	defer utils.DeferTx(tx, &err)
+	defer txutil.DeferTx(tx, &err)
 
 	err = s.Store.ModifyRestSchedule(ctx, tx, schedule)
 	if err != nil {
@@ -67,12 +68,12 @@ func (s *ServiceSchedule) ModifyRestSchedule(ctx context.Context, schedule entit
 // @param
 // -
 func (s *ServiceSchedule) RemoveRestSchedule(ctx context.Context, cno int64) (err error) {
-	tx, err := s.SafeTDB.BeginTx(ctx, nil)
+	tx, err := txutil.BeginTxWithMode(ctx, s.SafeTDB, false)
 	if err != nil {
 		return utils.CustomErrorf(err)
 	}
 
-	defer utils.DeferTx(tx, &err)
+	defer txutil.DeferTx(tx, &err)
 
 	err = s.Store.RemoveRestSchedule(ctx, tx, cno)
 	if err != nil {
