+++ csm-api/service/service_site.go
@@ -5,6 +5,7 @@ import (
 	"csm-api/auth"
 	"csm-api/entity"
 	"csm-api/store"
+	"csm-api/txutil"
 	"csm-api/utils"
 	"fmt"
 	"github.com/guregu/null"
@@ -187,12 +188,12 @@ func (s *ServiceSite) GetSiteStatsList(ctx context.Context, targetDate time.Time
 // @param
 // -
 func (s *ServiceSite) ModifySite(ctx context.Context, site entity.Site) (err error) {
-	tx, err := s.SafeTDB.BeginTx(ctx, nil)
+	tx, err := txutil.BeginTxWithMode(ctx, s.SafeTDB, false)
 	if err != nil {
 		return utils.CustomErrorf(err)
 	}
 
-	defer utils.DeferTx(tx, &err)
+	defer txutil.DeferTx(tx, &err)
 
 	if site.Sno.Int64 == 0 {
 		return utils.CustomErrorf(fmt.Errorf("sno parameter is missing"))
@@ -270,12 +271,12 @@ func (s *ServiceSite) ModifySite(ctx context.Context, site entity.Site) (err err
 // @param
 // -
 func (s *ServiceSite) AddSite(ctx context.Context, jno int64, user entity.User) (err error) {
-	tx, err := s.SafeTDB.BeginTx(ctx, nil)
+	tx, err := txutil.BeginTxWithMode(ctx, s.SafeTDB, false)
 	if err != nil {
 		return utils.CustomErrorf(err)
 	}
 
-	defer utils.DeferTx(tx, &err)
+	defer txutil.DeferTx(tx, &err)
 
 	err = s.Store.AddSite(ctx, s.SafeDB, tx, jno, user)
 	if err != nil {
@@ -289,12 +290,12 @@ func (s *ServiceSite) AddSite(ctx context.Context, jno int64, user entity.User) 
 // @param
 // -
 func (s *ServiceSite) ModifySiteIsNonUse(ctx context.Context, site entity.ReqSite) (err error) {
-	tx, err := s.SafeTDB.BeginTx(ctx, nil)
+	tx, err := txutil.BeginTxWithMode(ctx, s.SafeTDB, false)
 	if err != nil {
 		return utils.CustomErrorf(err)
 	}
 
-	defer utils.DeferTx(tx, &err)
+	defer txutil.DeferTx(tx, &err)
 
 	// 현장
 	if err = s.Store.ModifySiteIsNonUse(ctx, tx, site); err != nil {
@@ -322,12 +323,12 @@ func (s *ServiceSite) ModifySiteIsNonUse(ctx context.Context, site entity.ReqSit
 // @param
 // -
 func (s *ServiceSite) ModifySiteIsUse(ctx context.Context, site entity.ReqSite) (err error) {
-	tx, err := s.SafeTDB.BeginTx(ctx, nil)
+	tx, err := txutil.BeginTxWithMode(ctx, s.SafeTDB, false)
 	if err != nil {
 		return utils.CustomErrorf(err)
 	}
 
-	defer utils.DeferTx(tx, &err)
+	defer txutil.DeferTx(tx, &err)
 
 	// 현장
 	if err = s.Store.ModifySiteIsUse(ctx, tx, site); err != nil {
@@ -355,11 +356,12 @@ func (s *ServiceSite) ModifySiteIsUse(ctx context.Context, site entity.ReqSite) 
 // @param
 // -
 func (s *ServiceSite) SettingWorkRate(ctx context.Context, targetDate time.Time) (count int64, err error) {
-	tx, err := s.SafeTDB.BeginTx(ctx, nil)
+	tx, err := txutil.BeginTxWithMode(ctx, s.SafeTDB, false)
 	if err != nil {
 		return 0, utils.CustomErrorf(err)
 	}
-	defer utils.DeferTx(tx, &err)
+
+	defer txutil.DeferTx(tx, &err)
 
 	count, err = s.Store.SettingWorkRate(ctx, tx, targetDate)
 	if err != nil {
@@ -371,12 +373,12 @@ func (s *ServiceSite) SettingWorkRate(ctx context.Context, targetDate time.Time)
 
 // 공정률 수정
 func (s *ServiceSite) ModifyWorkRate(ctx context.Context, workRate entity.SiteWorkRate) (err error) {
-	tx, err := s.SafeTDB.BeginTx(ctx, nil)
+	tx, err := txutil.BeginTxWithMode(ctx, s.SafeTDB, false)
 	if err != nil {
 		return utils.CustomErrorf(err)
 	}
 
-	defer utils.DeferTx(tx, &err)
+	defer txutil.DeferTx(tx, &err)
 
 	err = s.Store.ModifyWorkRate(ctx, tx, workRate)
 	if err != nil {
