+++ csm-api/service/service_work_hour.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"csm-api/entity"
 	"csm-api/store"
+	"csm-api/txutil"
 	"csm-api/utils"
 )
 
@@ -15,12 +16,12 @@ type ServiceWorkHour struct {
 
 // 특정 프로젝트 및 근로자의 공수 계산: jno는 필수, ids는 없으면 jno의 모든 근로자 계산 있으면 해당 id의 근로자만 계산
 func (s *ServiceWorkHour) ModifyWorkHourByJno(ctx context.Context, jno int64, user entity.Base, ids []string) (err error) {
-	tx, err := s.SafeTDB.BeginTx(ctx, nil)
+	tx, err := txutil.BeginTxWithMode(ctx, s.SafeTDB, false)
 	if err != nil {
 		return utils.CustomErrorf(err)
 	}
 
-	defer utils.DeferTx(tx, &err)
+	defer txutil.DeferTx(tx, &err)
 
 	err = s.Store.ModifyWorkHourByJno(ctx, tx, jno, user, ids)
 	if err != nil {
@@ -31,12 +32,12 @@ func (s *ServiceWorkHour) ModifyWorkHourByJno(ctx context.Context, jno int64, us
 
 // 출퇴근이 둘다 있는 모든 근로자의 공수 계산
 func (s *ServiceWorkHour) ModifyWorkHour(ctx context.Context, user entity.Base) (err error) {
-	tx, err := s.SafeTDB.BeginTx(ctx, nil)
+	tx, err := txutil.BeginTxWithMode(ctx, s.SafeTDB, false)
 	if err != nil {
 		return utils.CustomErrorf(err)
 	}
 
-	defer utils.DeferTx(tx, &err)
+	defer txutil.DeferTx(tx, &err)
 
 	err = s.Store.ModifyWorkHour(ctx, tx, user)
 	if err != nil {
